+++ libs/coin-modules/coin-celo/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./types";
+export type { CeloTx, RLPEncodedTx } from "@celo/connect";
+export * from "@celo/wallet-base";