+++ libs/coin-modules/coin-mina/jest.config.js
@@ -5,4 +5,5 @@ module.exports = {
   testEnvironment: "node",
   testPathIgnorePatterns: ["lib/", "lib-es/", ".*\\.integ\\.test\\.[tj]s"],
   workerThreads: true,
+  setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };