+++ libs/coin-modules/coin-solana/src/signer.ts
@@ -23,4 +23,5 @@ export interface SolanaSigner {
     txBuffer: Buffer,
     resolution?: Resolution,
   ): Promise<SolanaSignature>;
+  signMessage(path: string, messageHex: string): Promise<SolanaSignature>;
 }