+++ libs/coin-tester-modules/coin-tester-evm/src/scenarii.test.ts
@@ -37,7 +37,8 @@ describe("EVM Deterministic Tester", () => {
     }
   });
 
-  it("scenario polygon", async () => {
+  // TODO: unskip when explorer is fixed on backend side
+  it.skip("scenario polygon", async () => {
     try {
       await executeScenario(scenarioPolygon);
     } catch (e) {