+++ libs/ledgerjs/packages/types-live/src/transaction.ts
@@ -50,6 +50,9 @@ export type SignOperationEvent = // Used when lot of exchange is needed with the
   | {
       type: "signed";
       signedOperation: SignedOperation;
+    }
+  | {
+      type: "transaction-checks-opt-in-triggered";
     };
 
 /**
@@ -71,6 +74,9 @@ export type SignOperationEventRaw =
   | {
       type: "signed";
       signedOperation: SignedOperationRaw;
+    }
+  | {
+      type: "transaction-checks-opt-in-triggered";
     };
 /**
  * Transaction is a generic object that holds all state for all transactions